+++ jest.config.ts
@@ -3,6 +3,7 @@ import './global-setup';
 module.exports = {
 	roots: ['<rootDir>'],
 	collectCoverage: false,
+	clearMocks: true,
 	coverageDirectory: 'coverage',
 	testEnvironment: 'node',
 	transform: {